WIP on support for '..' JSONPath operator.
The iter method of the '..' matcher will capture the tail of the path
and search for it.

The match method is trickier - seems nasty to have to pass in enough
context so the '..' match method can test the tail of the path.

This is to make it easier to have JSONTrigger trigger on events that
are rooted above an observed point but subtend a change that should
be observed.