Snapshot. Not entirely happy with the shape of it.
Is a segment an entity in a stream (which might be parsed from an
existing stream) or a concrete file? Possibly distinct subclasses - but
then where do things like encryption fit? And how do we map between
filenames and URLs. Sometimes we have no concrete file.

Probably overcomplicating something that would be simple as a
hacky script.